Support .and() in Reactive HttpBasic & HeaderBuilder
