+++ config/src/test/java/org/springframework/security/config/web/server/HeaderBuilderTests.java
@@ -42,7 +42,8 @@ import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
  * @since 5.0
  */
 public class HeaderBuilderTests {
-	HeaderBuilder headers = new HeaderBuilder();
+
+	HttpSecurity.HeaderBuilder headers = HttpSecurity.http().headers();
 
 	HttpHeaders expectedHeaders = new HttpHeaders();
